Import the latest Greentea client from Mbed OS
Development of the Greentea client was moved into Mbed OS, so this
repository was not updated for a long time. Having decided to separate
Greentea client as a standalone module, now we reuse this repository.
This PR brings in the latest changes.

Two important changes are
* Introduction of `greentea_getc()` and `greentea_putc()` to replace
direct usage of Mbed OS Serial API.
* Support for test metrics.

The imported files have been formatted using .astylerc from mbed-os.