cmake: Export greentea::client target
Exporting greentea-client CMake targets allows downstream projects to
discover and import the targets using find_package. After the targets
are imported users can reference them inside of their project. This
means that users won't be forced to vendor greentea-client if they
depend on it, they could instead install it to a known location where
CMake can look up the targets and import them.

The changes in this commit cause CMake to produce the necessary
information so greentea-client can easily be used by other CMake
projects from the local build directory, a local install or when
packaged.