Remove public version macro
With this macro, and version, checked into the repo, it also makes it
hard to acheive the ideal of "tag to release" (because the contents of
the repo would have to change at the same time as tagging). Mbed OS
didn't use this macro. Given the macro is likely unused, and we only use
it internally to share with the host what client version is running,
remove the version macro from the public API.