Obtain a version automatically
Attempt to determine a version automatically using git. Fallback to
v0.0.0 if the version can't be determined (e.g. due to git error, or if
the library is not from a git repo).

This makes it easier to tag-to-release as we don't have to edit a file
at the same time we make a tag: no repository contents need change when
making a release.