Add link to doc for the protocol sequence
As both the host and the device under test (DUT) are involved in the
Greentea protocol, the use of greentea-client and htrun need to be
explained in the same context. It's hard to break the explanation
into two, one for each sides. We leave the documentation in htrun
as the Greentea test sequence is largely host-driven.