cmake: Use `option` for user settable configuration
Previously we exposed an interface of cache variables allowing the user
to select different build configurations or build example applications.
However, it is more semantically correct to use `option` variables for
these configuration settings. Cache variables are usually understood to
be internal implementation details that can't be safely changed without
the risk of side effects. `option` variables are designed to be
user-facing, are limited to BOOL values, and are better supported in
ccmake and cmake-gui.