Full device (greentea-client) & host (htrun) demo
Add a demo which creates a PTY (pseudo-terminals) device node which
the host (htrun) can talk to like a serial device. To run it, two
shells are required: one for the device running greentea-client, one
for the host.

Notes:
* Given that htrun currently doesn't provide a generic way
(e.g. stdio) to interact with another process running greentea-client,
PTY is the easiest way to demo the full Greentea protocol without
requiring an embedded device or Mbed OS.
* PTY requires a Unix-like OS. Windows has a different API to access
terminals, so this example is excluded on Windows.