cmake: Use different targets to select IO types
Remove the option to select IO type and use another CMake target
instead. This allows more flexibility for users and means they can just
link to the target they need to build instead of trying to forward
option flags to greentea-client, which may not always be feasible if not
building the client from source.