Rename the examples
Previously we used the same name `greentea-client-example` for both
the stdio and custom_io examples, and relied on the switch
`GREENTEA_CLIENT_STDIO` to control which one gets built. But as
different examples, each one should have its own precise name for
clarity and to allow multiple examples to be built in the future.